Fix OOB read issue for Vdex 010
The current QuickeningInfo data for each method should be accessed only when
the iterator & code offset conditions are met. Otherwise, an OOB read happens
in the from the current_code_item_ptr. For most cases so far the OOB read was
not causing any issues since the value was never used. However, some Dex
layouts might trigger an un-mapped area SIGSEGV due to last code item being
close to the end of Vdex mapped area.

This patch moves the CurrentQuickeningInfo access inside the if statement to
prevent the OOB read.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>