[019] Better handling of deduplicated shared data sects
We need to append the deduplicated data section after the CompactDex main
section, otherwise the DexContainer is broken. In case of multi-dex files we pay
the size of cost of copying a very large data section to all exported files.
Hopefully, a better integration with the libdexlayout will minify back when
converting to StandardDex file.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>