Unhide APIs when decompiling files from Android P
As part of the API blacklisting in Android P, the Android runtime is implementing
a custom encoding in the access flags of the class members (fields & methods).
This is to prevent apps from directly messing with the core system frameworks.

However, when decompiling a framework Vdex the access flags that are written
on disk are encoded. So we need to decode them and overwrite when extracting
the Dex resources from the Vdex container. Otherwise, the ART verifier will
complain when the decompiled resources are repackaged or side-loaded.

This commit implements the base logic to decode access flags and the ULeb128
primitives to inline update the member values.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>