Improve handling of empty verified dependencies data
Upstream ART is implementing an unorthodox data population for the verified
dependencies. It always tries to '0' fill the size of all entries (9 bytes for 006 and 7
bytes for 010/019), although the header size is bigger. Therefore, the aggressive
boundary checks that were enforced, resulted into detected some Dex files with
empty data as malformed.

This commit fixes this bug by performing the vdeps data checks based on the
non-null offset, minimum expected length and number of entries per item.

Signed-off-by: Anestis Bechtsoudis <anestis@census-labs.com>